Split API app construction from startup
